# docs/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: NetDocGen API
  description: |
    Network Documentation Generator API
    
    NetDocGen transforms Visio network diagrams into professional documentation.
  version: 1.0.0
  contact:
    name: NetDocGen Support
    email: support@netdocgen.com
    url: https://netdocgen.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.netdocgen.com
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /api/collaboration/share:
    post:
      tags:
        - collaboration
      summary: Create share link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareLinkCreate'
      responses:
        '200':
          description: Share link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLink'
  /api/collaboration/comments:
    get:
      tags:
        - collaboration
      summary: List comments
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: document_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
    
    post:
      tags:
        - collaboration
      summary: Create comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '200':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ShareLinkCreate:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        permission:
          type: string
          enum: [view, edit, admin]
        expires_at:
          type: string
          format: date-time
        max_uses:
          type: integer
        password:
          type: string
    ShareLink:
      type: object
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
        share_url:
          type: string
        permission:
          type: string
          enum: [view, edit, admin]
        expires_at:
          type: string
          format: date-time
        use_count:
          type: integer
    CommentCreate:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 5000
        project_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        user_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time